+++ client/web/src/insights/core/backend/api/get-search-insight-content/query-has-count.test.ts
@@ -4,18 +4,21 @@ import { queryHasCountFilter } from './query-has-count-filter'
 
 describe('queryHasCount()', () => {
     it('returns true when count is specified', () => {
-        assert.strictEqual(queryHasCountFilter('const count: 1000'), true)
+        assert.strictEqual(queryHasCountFilter('const count:1000'), true)
     })
 
     it('returns false when count is not specified', () => {
         assert.strictEqual(queryHasCountFilter('const'), false)
     })
 
-    it('returns false when count is escaped', () => {
+    // Return these tests when Safari will support look behind group in regexp
+    // or when we have separate query parsers package
+    // See https://github.com/sourcegraph/sourcegraph/issues/22092
+    it.skip('returns false when count is escaped', () => {
         assert.strictEqual(queryHasCountFilter('"count:100" lang:ts'), false)
     })
 
-    it('returns false when count is escaped by single quotes', () => {
+    it.skip('returns false when count is escaped by single quotes', () => {
         assert.strictEqual(queryHasCountFilter("'count:100' lang:ts"), false)
     })
 })