+++ client/web/src/search/input/MonacoQueryInput.story.tsx
@@ -26,6 +26,7 @@ const defaultProps: MonacoQueryInputProps = {
     selectedSearchContextSpec: 'global',
     onChange: () => {},
     onSubmit: () => {},
+    settingsCascade: { final: null, subjects: null },
 }
 
 add(