+++ client/web/src/settings/tokens/AccessTokenCreatedAlert.tsx
@@ -1,5 +1,4 @@
 import classNames from 'classnames'
-import CheckmarkCircleIcon from 'mdi-react/CheckCircleIcon'
 import React from 'react'
 
 import { CodeSnippet } from '@sourcegraph/branded/src/components/CodeSnippet'
@@ -25,10 +24,7 @@ export const AccessTokenCreatedAlert: React.FunctionComponent<AccessTokenCreated
     const isSudoToken = token.scopes.includes(AccessTokenScopes.SiteAdminSudo)
     return (
         <div className={classNames('access-token-created-alert alert alert-success', className)}>
-            <p>
-                <CheckmarkCircleIcon className="redesign-d-none icon-inline" /> Copy the new access token now. You won't
-                be able to see it again.
-            </p>
+            <p>Copy the new access token now. You won't be able to see it again.</p>
             <CopyableText className="test-access-token" text={tokenSecret} size={48} />
             <h5 className="mt-4 mb-2">
                 <strong>Example usage</strong>