+++ docker-images/prometheus/cmd/prom-wrapper/mocks/gen.go
@@ -1,4 +1,3 @@
 package mocks
 
-//go:generate env GOBIN=$PWD/.bin GO111MODULE=on go install github.com/derision-test/go-mockgen/cmd/go-mockgen
-//go:generate $PWD/.bin/go-mockgen -f github.com/prometheus/client_golang/api/prometheus/v1 -i API -o prometheus_mock.go
+//go:generate ../../../../../dev/mockgen.sh github.com/prometheus/client_golang/api/prometheus/v1 -i API -o prometheus_mock.go